==> 3_extracting_outcomes.ipynb <==
# Extracting Outcomes of Decisions

- The following code will extract the outcomes of decisions from the processed text. 
- This involves identifying and isolating specific sections of the text that pertain to the decisions and their outcomes.

import pandas as pd
import regex as re

df = pd.read_csv(r'D:\Proyectos\db_comp\corpus.csv')

unique_values_count = df['Filename'].nunique()

print("Number of cases:", unique_values_count)

# For the purpose of this project, we will work only with case law written in English.

from langdetect import detect, LangDetectException

# Assuming df is already loaded and Text column is present
df['Text'] = df['Text'].fillna('')

# Function to detect if the text is in English
def is_english(text):
    try:
        return detect(text) == 'en'
    except LangDetectException:
        return False

# Apply the language detection function and classify as 'si' for English, 'no' otherwise
df['Classification'] = df['Text'].apply(lambda x: 'si' if is_english(x) else 'no')

en_case_law = df[df['Classification'] == 'si']
unique_values_count = en_case_law['Filename'].nunique()

print("Number of cases:", unique_values_count)

en_case_law.drop('Classification', axis=1, inplace=True)

# Keeping the rows that potentially contain outcome of the case law

pattern1 = r'D\s*E\s*C\s*I\s*S\s*I\s*O\s*N\s*:\s*'

# Use regex to check if either pattern appears in each row of the 'text' column
# Assign 'si' if either pattern is found, otherwise 'no'
en_case_law['Classification'] = en_case_law['Text'].apply(lambda x: 'si' if re.search(pattern1, x, flags=re.IGNORECASE) else 'no')

indices_to_keep = []

# Iterate over each unique filename
for filename in en_case_law['Filename'].unique():
    # Find the first occurrence of 'si' for the current filename
    first_si_index = en_case_law[(en_case_law['Filename'] == filename) & (en_case_law['Classification'] == 'si')].index.min()
    if pd.notnull(first_si_index):
        # Keep all rows after the first 'si' occurrence within the current filename
        indices_to_keep.extend(en_case_law.index[(en_case_law['Filename'] == filename) & (en_case_law.index >= first_si_index)].tolist())

# Filter the DataFrame using the indices to keep
final_decision = en_case_law.loc[indices_to_keep]

unique_values_count = final_decision['Filename'].nunique()

print("Number of unique values:", unique_values_count)

final_decision = final_decision.groupby('Filename')['Text'].apply(' '.join).reset_index()
final_decision['Text'] = final_decision['Text'].str.replace("\n", " ")

final_decision

cell_value = final_decision.at[1, 'Text']
cell_value

non_final_decisions = en_case_law[~en_case_law['Filename'].isin(final_decision['Filename'])]
non_final_decisions = non_final_decisions.groupby('Filename')['Text'].apply(' '.join).reset_index()
non_final_decisions['Text'] = non_final_decisions['Text'].str.replace("\n", " ")

final_decision.to_csv(r'D:\Proyectos\db_comp\en_final_decisions.csv')
non_final_decisions.to_csv(r'D:\Proyectos\db_comp\en_non_final_decisions.csv')

final_decision =pd.read_csv(r'D:\Proyectos\db_comp\en_final_decisions.csv')

pattern = r'D\s*E\s*C\s*I\s*S\s*I\s*O\s*N\s*:\s*'

# Function to extract text after pattern
def extract_text_after_pattern(text):
    match = re.search(pattern, text)
    if match:
        return text[match.end():]  # Extract text after the match
    else:
        return text

final_decision['Text'] = final_decision['Text'].apply(lambda x: extract_text_after_pattern(x))
final_decision

# First Article

# Define patterns
pattern1 = r'A\s*R\s*T\s*I\s*C\s*L\s*E\s*1\s*'
pattern2 = r'A\s*R\s*T\s*I\s*C\s*L\s*E\s*2\s*'

# Extract text using regex
def extract_text(row):
    match1 = re.search(pattern1, row['Text'], flags=re.IGNORECASE)
    match2 = re.search(pattern2, row['Text'], flags=re.IGNORECASE)

    if match1:
        start_index = match1.end()
    else:
        return ""

    if match2:
        end_index = match2.start()
    else:
        end_index = len(row['Text'])

    return row['Text'][start_index:end_index]

final_decision['Article 1'] = final_decision.apply(extract_text, axis=1)

# Second Article

# Define patterns
pattern1 = r'A\s*R\s*T\s*I\s*C\s*L\s*E\s*2\s*'
pattern2 = r'A\s*R\s*T\s*I\s*C\s*L\s*E\s*3\s*'

# Extract text using regex
def extract_text(row):
    match1 = re.search(pattern1, row['Text'], flags=re.IGNORECASE)
    match2 = re.search(pattern2, row['Text'], flags=re.IGNORECASE)

    if match1:
        start_index = match1.end()
    else:
        return ""

    if match2:
        end_index = match2.start()
    else:
        end_index = len(row['Text'])

    return row['Text'][start_index:end_index]

final_decision['Article 2'] = final_decision.apply(extract_text, axis=1)

# Third Article

# Define patterns
pattern1 = r'A\s*R\s*T\s*I\s*C\s*L\s*E\s*3\s*'
pattern2 = r'A\s*R\s*T\s*I\s*C\s*L\s*E\s*4\s*'

# Extract text using regex
def extract_text(row):
    match1 = re.search(pattern1, row['Text'], flags=re.IGNORECASE)
    match2 = re.search(pattern2, row['Text'], flags=re.IGNORECASE)

    if match1:
        start_index = match1.end()
    else:
        return ""

    if match2:
        end_index = match2.start()
    else:
        end_index = len(row['Text'])

    return row['Text'][start_index:end_index]

final_decision['Article 3'] = final_decision.apply(extract_text, axis=1)

final_decision.to_csv(r'D:\Proyectos\db_comp\en_outcomes.csv', index = False)
