==> 4_text_summarization.ipynb <==
## Text Summarization of Outcomes

This code includes a repository that performs text summarization of the extracted outcomes. It utilizes a zero-shot prompt with the GPT-3.5-turbo-0125 model. The model is prompted to determine whether the decision sanctioned the involved firms before summarizing.

### Prompt

Write a summary of the following outcome from a judicial decision. You must start with the word 'Sanction', if the judges found guilty, and then describe the conduct. If the judges did not find guilty, it must start with the word 'Not Sanction' and then describe the conduct.


import openai
from openai import OpenAI
client = OpenAI(api_key="Your API key")
import pandas as pd

final_decision = pd.read_csv(r'D:\Proyectos\db_comp\en_outcomes.csv')
final_decision

final_decision = final_decision.dropna(subset=['Article 1'])
final_decision

def generate_output(text):
    try:
        # Create chat completion
        completion = client.chat.completions.create(
            model="gpt-3.5-turbo-0125",
            messages=[
                {"role": "system", "content": "Write a summary of the following outcome from a judicial decision. You must start with the word 'Sanction', if the judges found guilty, and then describe the conduct. If the judges did not find guilty, it must start with the word 'Not Sanction' and then describe the conduct."},
                {"role": "user", "content": text}
            ]
        )
        return completion.choices[0].message.content
    except Exception as e:
        print(f"An error occurred: {e}")

final_decision['Summary'] = final_decision['Article 1'].apply(generate_output)
final_decision

final_decision.to_csv(r'D:\Proyectos\db_comp\en_classified_outcomes.csv', index=False)
