==> 5_creating_database.ipynb <==
## Merging Processed Data with Scraped Database

- This code includes merging the processed data with the scraped database. 

- It adds a column indicating whether the decision sanctioned or did not sanction the involved firms. 

- The final result is a database containing more than 500 decisions, each with a summary of the outcome.

import pandas as pd 

db_df = pd.read_csv(r'D:\Proyectos\db_comp\conducts_df.csv')
db_df

sum_df = pd.read_csv(r'D:\Proyectos\db_comp\en_classified_outcomes.csv')
sum_df['Filename'] = sum_df['Filename'].str.replace('.pdf', '')
sum_df

import numpy as np
sum_df['Sanction'] = np.where(sum_df['Summary'].str.contains('Not Sanction'),
                                         'no', 'yes')
sum_df

merged_df = pd.merge(sum_df, db_df, left_on='Filename', right_on='Title', how='inner')
merged_df = merged_df.drop_duplicates(subset=['Filename', 'Title'])

final_df = merged_df.drop(columns=['Unnamed: 0', 'Filename','Text', 'Article 1', 'Article 2', 'Article 3'])
final_df.to_excel(r'D:\Proyectos\db_comp\db_comp.xlsx', index=False)
